T#783: Extract 13 pack/beast routes to src/pack/routes.ts
Phase 2.5 of Library #102. Mechanical move, no logic changes.

Moved 13 routes: GET /api/pack, /api/pack/spinner-verbs, /api/beasts,
GET /api/beast/:name(/avatar.svg), PUT /api/beast/:name, PATCH /api/beast/:name(/avatar),
POST /api/beasts/seed-avatars, GET /api/beast/:name/terminal,
POST /api/beast/:name/terminal/input, POST /api/beast/:name/terminal/key,
POST /api/beast/:name/wake.

Helpers via DI: getTmuxStatus, normalizeAvatarUrl, webPresence,
WEB_PRESENCE_TIMEOUT_MS (server-internal state). Imports from db/index.ts
+ drizzle eq + node stdlib (fs, path, execSync).

Register call moved past L4902 (webPresence/WEB_PRESENCE_TIMEOUT_MS
declarations) to avoid TDZ error.

Server.ts: -366 lines. Cumulative T#771-T#783: -6,943 lines / -58%.
